Extract Decoder Dynamic tests
