+++ packages/evm/test/decoder/Decoder.spec.ts
@@ -3,8 +3,6 @@ import assert from "assert";
 import { expect } from "chai";
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
 
-import { AddressOne } from "@gnosis.pm/safe-contracts";
-import { BigNumber } from "ethers";
 import { defaultAbiCoder } from "ethers/lib/utils";
 
 import { Operator, ParameterType } from "../utils";
@@ -103,202 +101,6 @@ describe("Decoder library", async () => {
     await expect(
       decoder.pluck(data, result.children[1].location, result.children[1].size)
     ).to.be.reverted;
-  });
-
-  it("plucks Dynamic from top level", async () => {
-    const { decoder, testEncoder } = await loadFixture(setup);
-
-    const { data } =
-      await testEncoder.populateTransaction.dynamic32DynamicStatic(
-        [],
-        "Hello World!",
-        123456789
-      );
-
-    assert(data);
-
-    const layout = {
-      paramType: ParameterType.Calldata,
-      operator: Operator.Matches,
-      children: [
-        {
-          paramType: ParameterType.Array,
-          operator: Operator.Pass,
-          children: [
-            {
-              paramType: ParameterType.Static,
-              operator: Operator.Pass,
-              children: [],
-            },
-          ],
-        },
-        {
-          paramType: ParameterType.Dynamic,
-          operator: Operator.Pass,
-          children: [],
-        },
-        {
-          paramType: ParameterType.Static,
-          operator: Operator.Pass,
-          children: [],
-        },
-      ],
-    };
-
-    const result = await decoder.inspect(data, layout);
-
-    expect(
-      await decoder.pluck(
-        data,
-        result.children[1].location,
-        result.children[1].size
-      )
-    ).to.equal(encode(["string"], ["Hello World!"], YesRemoveOffset));
-  });
-  it("plucks Dynamic from Tuple", async () => {
-    const { decoder, testEncoder } = await loadFixture(setup);
-
-    const { data } = await testEncoder.populateTransaction._dynamicTuple({
-      dynamic: "0xabcd0011",
-    });
-
-    assert(data);
-    const layout = {
-      paramType: ParameterType.Calldata,
-      operator: Operator.Matches,
-      children: [
-        {
-          paramType: ParameterType.Tuple,
-          operator: Operator.Pass,
-          children: [
-            {
-              paramType: ParameterType.Dynamic,
-              operator: Operator.Pass,
-              children: [],
-            },
-          ],
-        },
-      ],
-    };
-
-    const result = await decoder.inspect(data, layout);
-
-    expect(
-      await decoder.pluck(
-        data,
-        result.children[0].location,
-        result.children[0].size
-      )
-    ).to.equal(encode(["tuple(bytes)"], [["0xabcd0011"]], YesRemoveOffset));
-  });
-  it("plucks Dynamic from Array", async () => {
-    const { decoder, testEncoder } = await loadFixture(setup);
-
-    const { data } = await testEncoder.populateTransaction.dynamicArray([
-      "0xaabbccdd",
-      "0x004466ff",
-    ]);
-
-    const layout = {
-      paramType: ParameterType.Calldata,
-      operator: Operator.Matches,
-      children: [
-        {
-          paramType: ParameterType.Array,
-          operator: Operator.Pass,
-          children: [
-            {
-              paramType: ParameterType.Dynamic,
-              operator: Operator.Pass,
-              children: [],
-            },
-          ],
-        },
-      ],
-    };
-
-    const result = await decoder.inspect(data as string, layout);
-
-    const arrayElement0 = result.children[0].children[0];
-    const arrayElement1 = result.children[0].children[1];
-    expect(
-      await decoder.pluck(
-        data as string,
-        arrayElement0.location,
-        arrayElement0.size
-      )
-    ).to.equal(encode(["bytes"], ["0xaabbccdd"], YesRemoveOffset));
-
-    expect(
-      await decoder.pluck(
-        data as string,
-        arrayElement1.location,
-        arrayElement1.size
-      )
-    ).to.equal(encode(["bytes"], ["0x004466ff"], YesRemoveOffset));
-  });
-  it("plucks Dynamic from nested AbiEncoded", async () => {
-    const { decoder, testEncoder } = await loadFixture(setup);
-
-    const { data: nestedData } =
-      await testEncoder.populateTransaction.dynamicStaticDynamic32(
-        "0xbadfed",
-        true,
-        ["0xccdd"]
-      );
-
-    assert(nestedData);
-
-    const { data } = await testEncoder.populateTransaction._dynamicTuple({
-      dynamic: nestedData,
-    });
-
-    const nestedLayout = {
-      paramType: ParameterType.Calldata,
-      operator: Operator.Matches,
-      children: [
-        {
-          paramType: ParameterType.Dynamic,
-          operator: Operator.Pass,
-          children: [],
-        },
-        {
-          paramType: ParameterType.Static,
-          operator: Operator.Pass,
-          children: [],
-        },
-        {
-          paramType: ParameterType.Array,
-          operator: Operator.Pass,
-          children: [
-            {
-              paramType: ParameterType.Static,
-              operator: Operator.Pass,
-              children: [],
-            },
-          ],
-        },
-      ],
-    };
-
-    const layout = {
-      paramType: ParameterType.Calldata,
-      operator: Operator.Matches,
-      children: [
-        {
-          paramType: ParameterType.Tuple,
-          operator: Operator.Pass,
-          children: [nestedLayout],
-        },
-      ],
-    };
-
-    const result = await decoder.inspect(data as string, layout);
-
-    const tupleField = result.children[0].children[0].children[0];
-    expect(
-      await decoder.pluck(data as string, tupleField.location, tupleField.size)
-    ).to.equal(encode(["bytes"], ["0xbadfed"], YesRemoveOffset));
   });
 
   it.skip("plucks Tuple from top level");
